feat: IPv6 support

+++ include/raknet/RakPeer.h
@@ -137,7 +137,7 @@ public:
     ///
     /// Defaults to 0, meaning by default, nobody can connect to you
     /// \param[in] numberAllowed Maximum number of incoming connections allowed.
-    void SetMaximumIncomingConnections(unsigned short numberAllowed);
+    void SetMaximumIncomingConnections(uint32_t numberAllowed);
 
     /// \brief Returns the value passed to SetMaximumIncomingConnections().
     /// \return Maximum number of incoming connections, which is always <= maxConnections
